requirements.txt
django-endless-pagination==1.1